requirements.txt
requests==2.31.0
msal==1.24.1
python-dotenv==1.0.0

